requirements.txt
streamlit
transformers
torch
nltk
plotly
gtts
textstat
python-docx
PyMuPDF
google-generativeai
language-tool-python
